+++ src/components/navigation/Navbar.tsx
@@ -2,7 +2,6 @@
 import {
   Box,
   BoxProps,
-  Container,
   Drawer,
   Link,
   LinkProps,
@@ -25,6 +24,14 @@ interface NavLinksProps {
 interface NavbarProps extends BoxProps, React.PropsWithChildren {
   logo?: ImageColorSchemeSwitchType | "theme" | null;
   linkComponent?: React.ElementType;
+  centreSlot?: React.ReactElement<LinkProps> | React.ReactElement<LinkProps>[];
+  rightSlot?: React.ReactElement<LinkProps> | React.ReactElement<LinkProps>[];
+  logoLeftSlot?:
+    | React.ReactElement<LinkProps>
+    | React.ReactElement<LinkProps>[];
+  logoRightSlot?:
+    | React.ReactElement<LinkProps>
+    | React.ReactElement<LinkProps>[];
 }
 
 interface NavLinkProps extends LinkProps {
@@ -147,7 +154,16 @@ const BoxStyled = styled(Box)<BoxProps>(({ theme }) => ({
 /**
  * Basic navigation bar. Can be used with `NavLinks` and `NavLink` to display a responsive list of links.
  */
-const Navbar = ({ children, logo, linkComponent, ...props }: NavbarProps) => {
+const Navbar = ({
+  children,
+  logo,
+  linkComponent,
+  logoLeftSlot,
+  logoRightSlot,
+  rightSlot,
+  centreSlot,
+  ...props
+}: NavbarProps) => {
   const theme = useTheme();
   let resolvedLogo: ImageColorSchemeSwitchType | null | undefined = null;
   if (logo === "theme") {
@@ -158,12 +174,19 @@ const Navbar = ({ children, logo, linkComponent, ...props }: NavbarProps) => {
 
   return (
     <BoxStyled role="banner" {...props}>
-      <Container maxWidth="lg" sx={{ height: "100%" }}>
-        <Stack
-          direction="row"
-          spacing={8}
-          sx={{ height: "100%", alignItems: "center", width: "100%" }}
-        >
+      <Stack
+        justifyContent="space-between"
+        direction="row"
+        spacing={8}
+        margin={2}
+        sx={{
+          height: "100%",
+          alignItems: "center",
+          width: "100%",
+        }}
+      >
+        <Stack direction="row" alignItems="center" spacing={2}>
+          {logoLeftSlot}
           {resolvedLogo && (
             <Link
               key="logo"
@@ -182,9 +205,21 @@ const Navbar = ({ children, logo, linkComponent, ...props }: NavbarProps) => {
               </Box>
             </Link>
           )}
+          {logoRightSlot}
           {children}
         </Stack>
-      </Container>
+
+        {rightSlot}
+      </Stack>
+      <Box
+        sx={{
+          position: "absolute",
+          left: "50%",
+          transform: "translateX(-50%)",
+        }}
+      >
+        {centreSlot}
+      </Box>
     </BoxStyled>
   );
 };