+++ changelog.md
@@ -14,6 +14,7 @@ SciReactUI Changelog
 ### Changed
 - Breadcrumbs component takes optional linkComponent prop for page routing. 
 - Navbar, NavLink and FooterLink will use routing library for links if provided with linkComponent and to props.
+- Navbar uses slots for positioning elements. Breaking change: elements must now use rightSlot for positioning to the far right.
 
 [v0.1.0] - 2025-04-10
 ---------------------